requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.1.3
blinker==1.8.2
boilerpy3==1.0.7
boto3==1.34.116
botocore==1.34.116
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-haystack==0.17.0
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
clickhouse-connect==0.7.11
coloredlogs==15.0.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
duckdb==0.10.3
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.99.1
fastapi-cli==0.0.4
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fsspec==2024.5.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
graphlib_backport==1.1.0
grpcio==1.64.0
h11==0.14.0
haystack-ai==2.1.2
haystack-bm25==1.0.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kubernetes==29.0.0
lazy-imports==0.3.1
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.30.5
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pluggy==1.5.0
posthog==3.5.0
protobuf==4.25.3
pulsar-client==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydantic==1.10.15
pydantic_core==2.18.3
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pysqlite3-binary==0.5.2.post3
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.12.1
tabulate==0.9.0
tenacity==8.3.0
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.18
uvicorn==0.30.0
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.0
zstandard==0.22.0
